+++ src/main/java/net/silentchaos512/mechanisms/block/alloysmelter/AlloySmelterScreen.java
@@ -1,0 +1,30 @@
+package net.silentchaos512.mechanisms.block.alloysmelter;
+
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.text.ITextComponent;
+import net.silentchaos512.mechanisms.SilentMechanisms;
+import net.silentchaos512.mechanisms.block.AbstractMachineScreen;
+
+public class AlloySmelterScreen extends AbstractMachineScreen<AlloySmelterContainer> {
+    public static final ResourceLocation TEXTURE = SilentMechanisms.getId("textures/gui/alloy_smelter.png");
+
+    public AlloySmelterScreen(AlloySmelterContainer containerIn, PlayerInventory playerInventoryIn, ITextComponent titleIn) {
+        super(containerIn, playerInventoryIn, titleIn);
+    }
+
+    @Override
+    public ResourceLocation getGuiTexture() {
+        return TEXTURE;
+    }
+
+    @Override
+    protected int getProgressArrowPosX(int guiPosX) {
+        return guiPosX + 92;
+    }
+
+    @Override
+    protected int getProgressArrowPosY(int guiPosY) {
+        return guiPosY + 35;
+    }
+}