+++ src/main/java/net/silentchaos512/mechanisms/init/ModTileEntities.java
@@ -6,6 +6,7 @@ import net.minecraft.tileentity.TileEntityType;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.silentchaos512.mechanisms.SilentMechanisms;
+import net.silentchaos512.mechanisms.block.alloysmelter.AlloySmelterTileEntity;
 import net.silentchaos512.mechanisms.block.batterybox.BatteryBoxTileEntity;
 import net.silentchaos512.mechanisms.block.compressor.CompressorTileEntity;
 import net.silentchaos512.mechanisms.block.crusher.CrusherTileEntity;
@@ -16,6 +17,7 @@ import net.silentchaos512.mechanisms.block.wire.WireTileEntity;
 import java.util.function.Supplier;
 
 public class ModTileEntities {
+    public static TileEntityType<AlloySmelterTileEntity> alloySmelter;
     public static TileEntityType<BatteryBoxTileEntity> batteryBox;
     public static TileEntityType<CoalGeneratorTileEntity> coalGenerator;
     public static TileEntityType<CompressorTileEntity> compressor;
@@ -24,6 +26,7 @@ public class ModTileEntities {
     public static TileEntityType<WireTileEntity> wire;
 
     public static void registerAll(RegistryEvent.Register<TileEntityType<?>> event) {
+        alloySmelter = register("alloy_smelter", AlloySmelterTileEntity::new, ModBlocks.alloySmelter);
         batteryBox = register("battery_box", BatteryBoxTileEntity::new, ModBlocks.batteryBox);
         coalGenerator = register("coal_generator", CoalGeneratorTileEntity::new, ModBlocks.coalGenerator);
         compressor = register("compressor", CompressorTileEntity::new, ModBlocks.compressor);