+++ src/main/java/net/silentchaos512/mechanisms/util/Constants.java
@@ -4,6 +4,7 @@ import net.minecraft.util.ResourceLocation;
 import net.silentchaos512.mechanisms.SilentMechanisms;
 
 public final class Constants {
+    public static final ResourceLocation ALLOY_SMELTING = SilentMechanisms.getId("alloy_smelting");
     public static final ResourceLocation COMPRESSING = SilentMechanisms.getId("compressing");
     public static final ResourceLocation CRUSHING = SilentMechanisms.getId("crushing");